% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{baseline}
\alias{baseline}
\title{DE results for airway - baseline}
\format{
A DESEqResults
}
\value{
A DESEqResults
}
\description{
Example DE results for airway data set
}
\details{
This is a \code{DESEqResults} object, with the simplest model, using \code{~dex} as
the design. Data is from the airway package, DE results are generated with \code{DESeq2}.
}
