% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{cell_controlled}
\alias{cell_controlled}
\title{DE results for airway - cell-controlled}
\format{
A DESEqResults
}
\value{
A DESEqResults
}
\description{
Example DE results for airway data set
}
\details{
This is a \code{DESEqResults} object, using \code{~ cell + dex} as the design.
Data is from the airway package, DE results are generated with \code{DESeq2},
using the \verb{dex untrt vs trt} contrast
}
