% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{controlled_down_go}
\alias{controlled_down_go}
\title{Example GO enrichments for the cell_controlled model}
\format{
A EnrichResults
}
\value{
A EnrichResults object
}
\description{
GO enrichments based on the \code{cell_controlled} DE results
}
\details{
GO enrichmens are generated with \code{clusterProfiler}, filtered on down-regulated
genes only; genes that underwent DE testing were taken as the universe.

The code to create said object can be found in the folder \verb{/inst/scripts} in
the exploreSE package, the file is called \code{creating_dataset_examples.R}.
}
