% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreSE-package.R
\docType{package}
\name{exploreSE-package}
\alias{exploreSE-package}
\title{exploreSE: Explorer for Your Bulk RNA Seq Analyses}
\description{
Interactive exploration of a Summarized experiment and associated DE analyses. A shiny app to inspect and interact with pre-calculated differential expression and functional enrichment results, and to facilitate comparisons between multiple models. This is meant to be a tool for exploration and discussion, rather than analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jaspitzer/exploreSE}
  \item Report bugs at \url{https://github.com/jaspitzer/exploreSE/issues}
}

}
\author{
\strong{Maintainer}: Jasper Spitzer \email{jasspitzer135@gmail.com} (\href{https://orcid.org/0000-0001-9696-2092}{ORCID})

Authors:
\itemize{
  \item Jasper Spitzer \email{jasspitzer135@gmail.com} (\href{https://orcid.org/0000-0001-9696-2092}{ORCID})
}

}
\keyword{internal}
