% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_utils.R
\name{get.gos}
\alias{get.gos}
\title{Calculate GO Enrichments for a given comparison}
\usage{
get.gos(NAME, obj, species = "hs", gene_type = "SYMBOL")
}
\arguments{
\item{NAME}{character, name of the DE comparison}

\item{obj}{DeeDeeExperiment or SummarizedExperiment, the object}

\item{species}{character, the species hs/mm}

\item{gene_type}{character, the type of gene identifier}
}
\value{
a DeeDeeExperiment or SummarizedExperiment
}
\description{
A function to calculate GO term enrichments from a given summarizedExperiment
or DeeDee experiment. By default, it calculates separate enrichments for
up- downregulated genes, at the cutoffs of |logFC| > 1 and adjusted p-
value < 0.05. It utilises the \code{\link[clusterProfiler:enrichGO]{clusterProfiler::enrichGO()}} function.
}
\examples{

library(airway)
library(DeeDeeExperiment)
data(airway)
data("baseline_de", package = "exploreSE")

se_with_de <- addDEA(DeeDeeExperiment(airway), baseline)
se_with_go <- get.gos(obj = se_with_de,
                      NAME = "baseline",
                      gene_type = "ENSEMBL")

}
