% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{se_with_de}
\alias{se_with_de}
\title{An example data set with added DE results}
\format{
A DeeDeeExperiment
}
\value{
A DeeDeeExperiment with two DEA analyses
}
\description{
An example data set with added DE results. Data is from the airway package,
DE results are generated with \code{DESeq2}.
}
\details{
This is a \code{DeeDeeExperiment} object, a derivative of the airway data set.
In the DEA slot are two separate differential expression analyses, both
generated with \code{DESeq2}.
}
