% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{se_with_go}
\alias{se_with_go}
\title{An example data set with added FE results}
\format{
A DeeDeeExperiment
}
\value{
A DeeDeeExperiment with two FEA analyses
}
\description{
An example data set with added FE results. Data is from the airway package,
DE results are generated with \code{clusterProfiler}.
}
\details{
This is a \code{DeeDeeExperiment} object, a derivative of the airway data set.
In the DEA slot are two separate functional enrichment analyses, both
generated with \code{clusterProfiler}.
}
