% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreSE.R
\name{exploreSE}
\alias{exploreSE}
\title{exploreSE}
\usage{
exploreSE()
}
\value{
an application
}
\description{
this runs the explorer app
}
\examples{
app <- exploreSE()
if(interactive()){
shiny::runApp(app, port = 1234)
}
}
