% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_utils.R
\name{get.gos}
\alias{get.gos}
\title{Title}
\usage{
get.gos(NAME, obj = dds, species = "hs", gene_type = "SYMBOL")
}
\arguments{
\item{NAME}{name of the de comparison}

\item{obj}{the object}

\item{species}{the species hs/mm}

\item{gene_type}{the type of gene identifier}
}
\value{
a DeeDeeExperiment or SummarizedExperiment
}
\description{
Title
}
\examples{
TRUE
}
