% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_utils.R
\name{get.gsea}
\alias{get.gsea}
\title{Title}
\usage{
get.gsea(
  NAME,
  obj = dds,
  type = "HALLMARK",
  conditions,
  species = "hs",
  condition_var = condition
)
}
\arguments{
\item{NAME}{name of the DE comparison}

\item{obj}{the SE/DeeDeeExperiemnt}

\item{type}{hallmark or reactome}

\item{conditions}{which conditions are compared}

\item{species}{what species}

\item{condition_var}{whats the variable name for this comparison}
}
\value{
a DeeDeeExperiment or SummarizedExperiment
}
\description{
Title
}
\examples{
TRUE
}
